Split the pydantic-ai harness along the four jobs it was doing
`pydantic_ai.py` was 1293 lines and `_execute` scored 42 on the complexity
ratchet -- the highest number in the backend. It was doing four separable jobs
inside one function: building the agent, walking the run graph, streaming each
model request into events, and retrying a dropped stream.

* `pydantic_ai_streaming` -- one model request into the tokens a user watches
  appear. `agent_run_id` and `should_stop` became fields, which is worth more
  than it sounds: the stop check appeared eleven times as
  `if await self._should_stop(should_stop): yield self._stopped_event(agent_run_id)`,
  with both values threaded through six signatures to reach it.
* `pydantic_ai_node_loop` -- one pass over the run graph, with the two rules
  that keep a retry safe stated at the top: a model response is persisted
  all-or-nothing, and the snapshot must include `node.request` or pydantic-ai
  re-runs tools that already ran.
* `pydantic_ai_retry` -- the retry policy, and why a relayed `CancelledError` is
  only ours to swallow when we are the ones who cancelled.
* `pydantic_ai_usage` -- carrying token counts across abandoned attempts, which
  the provider still bills for.

`_execute` is 18 and the file is 542. The anyio invariant is unchanged and now
stated in the node loop's own docstring: the whole loop runs in a child task
because `agent.iter()` enters cancel scopes bound to their creating task, and
unwinding them anywhere else crashes the worker.

Also in this pass, three functions the ratchet had been holding flat:

* `_user_prompt_text` (32) is one block per kind of context -- sender, channel,
  attachments, reply rules. The channel block carries a security-shaped comment
  about not acting on other participants' messages; it is now a function with
  that comment as its docstring rather than a paragraph in the middle of a
  pipeline. Two of its three broad catches narrowed to `ImportError`, which is
  what a lazy import actually fails with.
* `_update_provider` (41) had patch resolution interleaved with SSRF validation,
  rediscovery and the write. `resolve_provider_patch` is now pure -- no DB, no
  network -- so the three-way meaning of an optional field (absent keeps, a
  string replaces, null clears) is testable without either.
* `validate_display_payload` (29) is one validator per resource type. It
  validates model-authored payloads and its return value is what the model reads
  to fix its own call, so a single condition covering four types could only say
  something vague -- and a vague error is one the model retries verbatim.

The pod toolset split tables from files, which takes the last agent file off the
oversized list: **zero oversized files in the module**, down from seven.
