+++ lemma-backend/app/modules/agent/infrastructure/harnesses/pydantic_ai.py
@@ -3,8 +3,6 @@
 from __future__ import annotations
 
 import asyncio
-import json
-import random
 from collections.abc import Awaitable, Callable
 from typing import AsyncIterator, Sequence
 from uuid import UUID
@@ -16,24 +14,22 @@ from app.modules.agent.infrastructure.harnesses.mock_model import (
     is_mock_llm_enabled,
 )
 from pydantic_ai import Agent as PydanticAIAgent
-from pydantic_ai import BinaryContent, FunctionToolCallEvent, FunctionToolResultEvent
 from pydantic_ai.exceptions import (
     ModelHTTPError,
     UnexpectedModelBehavior,
     UsageLimitExceeded,
 )
 from pydantic_ai.messages import (
-    FinalResultEvent,
     ModelMessage,
-    PartDeltaEvent,
-    PartEndEvent,
-    PartStartEvent,
-    ThinkingPart,
-    ThinkingPartDelta,
-    TextPart,
-    TextPartDelta,
-    ToolCallPart,
-    ToolCallPartDelta,
+)
+from app.modules.agent.infrastructure.harnesses.pydantic_ai_node_loop import NodeLoop
+from app.modules.agent.infrastructure.harnesses.pydantic_ai_retry import (
+    HarnessDriverCancelled,
+    _drive_with_retry,
+    _reraise_driver_failure,
+)
+from app.modules.agent.infrastructure.harnesses.pydantic_ai_streaming import (
+    ModelRequestStreamer,
 )
 from app.modules.agent.domain.context import AgentContext
 from app.modules.agent.domain.entities import Agent, Conversation, Message
@@ -42,7 +38,6 @@ from app.modules.agent.services.run_phase_spans import run_phase
 from app.modules.agent.domain.value_objects import (
     AgentEvent,
     AgentEventType,
-    AgentRunUsage,
     HarnessKind,
     HarnessOptions,
     JsonObject,
@@ -58,12 +53,8 @@ from app.modules.agent.infrastructure.harnesses.pydantic_ai_history import (
 from app.modules.agent.infrastructure.harnesses.provider_error_log import (
     log_model_http_error,
 )
-from app.modules.agent.infrastructure.harnesses.pydantic_ai_stream_parts import (
-    StreamingParts,
-)
 from app.modules.agent.infrastructure.transport_errors import (
     is_retryable_stream_error,
-    retry_after_seconds,
 )
 from app.modules.agent.config import agent_settings
 from app.modules.agent.services.runtime_model_factory import (
@@ -80,18 +71,6 @@ from app.core.request_context import create_inherited_task
 
 logger = get_logger(__name__)
 StopChecker = Callable[[], Awaitable[bool]]
-
-
-class HarnessDriverCancelled(Exception):
-    """The graph-driving task was cancelled while the run was still healthy.
-
-    An ordinary `Exception` rather than a `CancelledError` subclass on purpose:
-    nothing about *this* task is being cancelled, so every `except
-    CancelledError` between here and the runner would draw the wrong conclusion
-    and re-raise a cancellation that isn't happening. Raised as a failure
-    because that is what it is — the graph stopped mid-node and whatever tool
-    was executing was cancelled with it.
-    """
 
 
 def _user_facing_error_message(exc: Exception) -> str:
@@ -166,7 +145,6 @@ def _user_facing_error_message(exc: Exception) -> str:
 DEFAULT_TOOL_RETRIES = 5
 
 # Ceiling for the pause between stream-drop retries; see `_retry_backoff`.
-_RETRY_BACKOFF_CAP_SECONDS = 6.0
 
 
 def _instructions(
@@ -190,112 +168,6 @@ def _instructions(
         )
         span.set_attribute("lemma.instructions.chars", len(text))
         return text
-
-
-def _reraise_driver_failure(
-    pending_error: BaseException | None,
-    *,
-    cancelled_by_us: bool,
-    agent_run_id: UUID,
-) -> None:
-    """Decide what a failure relayed out of the driver task means.
-
-    A real error is re-raised so ``run()``'s handlers (ModelHTTPError,
-    UsageLimitExceeded, AgentInputRequired, …) still fire.
-
-    A relayed ``CancelledError`` is only ours to drop when we are the ones who
-    cancelled — either our own cancellation is already propagating out of the
-    consumer loop, or we tore the driver down on the way out. Otherwise the
-    driver died under a healthy parent: the graph stopped mid-node, so any tool
-    still executing was cancelled with it and its call will never get a result.
-
-    Dropping that silently is how a truncated run came to report success. The
-    generator returned normally, so ``run()`` emitted COMPLETED and the run was
-    finalized with no error, no log line, and a trailing tool call that nothing
-    ever answered — the agent simply stopped, and every layer above said it went
-    fine.
-    """
-
-    if pending_error is None:
-        return
-    if not isinstance(pending_error, asyncio.CancelledError):
-        raise pending_error
-    if cancelled_by_us:
-        return
-    logger.error(
-        "agent.pydantic_ai.driver_cancelled_mid_run.failed",
-        agent_run_id=str(agent_run_id),
-        exc_info=pending_error,
-    )
-    raise HarnessDriverCancelled(
-        "The agent run was cancelled while a tool call was still running."
-    ) from pending_error
-
-
-async def _drive_with_retry(
-    drive_once,
-    *,
-    queue,
-    max_attempts: int,
-    stream_reset,
-    stopped,
-    should_stop,
-) -> None:
-    """Run one agent-graph pass, re-entering it after a transient stream drop.
-
-    Lives at module level rather than nested in `_execute` so the retry policy
-    can be read — and reasoned about — without the several hundred lines of
-    streaming machinery it sits inside.
-
-    `drive_once` publishes the live run and a pre-request history snapshot into
-    the `state` dict it is handed. Both are required to resume: the run carries
-    the usage already spent, and the snapshot is the history as it stood
-    *before* the failing request, which is what keeps a retry from replaying a
-    half-written response or re-executing a completed tool.
-    """
-    carried_usage: dict[str, int] = {}
-    resume_history: list[object] | None = None
-    attempt = 0
-    while True:
-        state: dict[str, object] = {}
-        try:
-            await drive_once(resume_history, state, carried_usage)
-            return
-        # Deliberately `Exception`, not `BaseException`: a CancelledError must
-        # reach the caller untouched so pydantic-graph's scopes unwind in this
-        # task. It would have been re-raised here anyway —
-        # `is_retryable_stream_error` never retries cancellation — so letting it
-        # skip the handler entirely is the same behaviour with less to reason
-        # about.
-        except Exception as exc:  # noqa: BLE001 — re-raised below when fatal
-            run = state.get("run")
-            snapshot = state.get("resume_from")
-            if (
-                run is None
-                or snapshot is None
-                or attempt + 1 >= max_attempts
-                or not is_retryable_stream_error(exc)
-            ):
-                raise
-            # Resuming re-asks only the request that failed. Completed tool
-            # results are replayed from the snapshot rather than re-executed,
-            # so no tool ever runs twice.
-            _accumulate_usage(carried_usage, getattr(run, "usage", None))
-            resume_history = list(snapshot)  # type: ignore[arg-type]
-            attempt += 1
-            logger.warning(
-                "agent.pydantic_ai.model_stream_retry.degraded",
-                attempt=attempt,
-                max_attempts=max_attempts,
-                error_type=type(exc).__name__,
-            )
-            # Tell the client to drop the half-streamed bubble; the replacement
-            # response streams from scratch.
-            await queue.put(("event", stream_reset()))
-            if await should_stop():
-                await queue.put(("event", stopped()))
-                return
-            await asyncio.sleep(retry_after_seconds(exc) or _retry_backoff(attempt))
 
 
 class PydanticAIHarness:
@@ -522,200 +394,31 @@ class PydanticAIHarness:
         # cancel scope". Events stream out through a queue.
         queue: asyncio.Queue[tuple[str, object]] = asyncio.Queue()
         max_stream_attempts = max(1, agent_settings.agent_model_stream_max_attempts)
+        streamer = ModelRequestStreamer(
+            emit_tokens=self.emit_tokens,
+            agent_run_id=agent_run_id,
+            should_stop=should_stop,
+        )
 
-        async def _drive_once(
-            resume_history: list[object] | None,
-            state: dict[str, object],
-            carried_usage: dict[str, int],
-        ) -> None:
-            async with _new_run_context(resume_history) as run:
-                # Published before the node loop so the retry handler can read
-                # usage off the run that just failed.
-                state["run"] = run
-                async for node in run:
-                    if PydanticAIAgent.is_model_request_node(node):
-                        # Snapshot BEFORE streaming, because after a
-                        # mid-stream failure `all_messages()` has grown a
-                        # truncated ModelResponse plus an empty
-                        # ModelRequest: resuming from that makes the model
-                        # continue an answer whose first half we threw away.
-                        #
-                        # `node.request` must be appended explicitly. It is
-                        # the request about to be sent (for a post-tool node,
-                        # the one carrying the ToolReturnParts) and it does
-                        # not enter `all_messages()` until the request
-                        # actually happens. Without it the resume history
-                        # ends on a ModelResponse whose tool calls have no
-                        # results, and pydantic-ai — correctly — executes
-                        # those tools again. That is how a retry would
-                        # double-charge a card.
-                        snapshot = list(run.all_messages())
-                        pending = getattr(node, "request", None)
-                        if pending is not None and (
-                            not snapshot or snapshot[-1] is not pending
-                        ):
-                            snapshot.append(pending)
-                        state["resume_from"] = snapshot
-                        # A model response is persisted all-or-nothing. Parts
-                        # that finish before the stream dies are NOT in
-                        # `run.all_messages()` (pydantic-ai appends the
-                        # ModelResponse only once the node completes), so
-                        # writing them as they arrive would duplicate them on
-                        # resume. Tokens still stream live; only the durable
-                        # write waits for the node to finish.
-                        buffered: list[AgentEvent] = []
-                        async for event in self._stream_model_request(
-                            node,
-                            run,
-                            agent_run_id=agent_run_id,
-                            malformed_tool_call_ids=malformed_tool_call_ids,
-                            should_stop=should_stop,
-                        ):
-                            if event.type is AgentEventType.MESSAGE:
-                                buffered.append(event)
-                                continue
-                            if event.type in {
-                                AgentEventType.ERROR,
-                                AgentEventType.STOPPED,
-                            }:
-                                # A stop is a real ending, not a lost response:
-                                # flush what the model already produced before
-                                # the terminal event so the user keeps it.
-                                for held in buffered:
-                                    await queue.put(("event", held))
-                                await queue.put(("event", event))
-                                return
-                            await queue.put(("event", event))
-                        for held in buffered:
-                            await queue.put(("event", held))
-                    elif PydanticAIAgent.is_call_tools_node(node):
-                        async for event in self._stream_tool_calls(
-                            node,
-                            run,
-                            conversation_id=ctx.conversation_id,
-                            agent_run_id=agent_run_id,
-                            malformed_tool_call_ids=malformed_tool_call_ids,
-                            emitted_tool_response_ids=emitted_tool_response_ids,
-                            should_stop=should_stop,
-                        ):
-                            await queue.put(("event", event))
-                            if event.type in {
-                                AgentEventType.ERROR,
-                                AgentEventType.STOPPED,
-                            }:
-                                return
-                    elif PydanticAIAgent.is_end_node(node):
-                        if node.data.tool_call_id:
-                            if node.data.tool_call_id not in emitted_tool_response_ids:
-                                await queue.put(
-                                    (
-                                        "event",
-                                        AgentEvent(
-                                            type=AgentEventType.MESSAGE,
-                                            data=MessageDraft.of_tool_return(
-                                                tool_name=node.data.tool_name
-                                                or "unknown_tool",
-                                                tool_call_id=node.data.tool_call_id,
-                                                tool_result=to_json_value(
-                                                    node.data.output
-                                                ),
-                                                metadata={
-                                                    "tool_name": node.data.tool_name
-                                                    or "unknown_tool"
-                                                },
-                                            ),
-                                            agent_run_id=agent_run_id,
-                                        ),
-                                    )
-                                )
-                                if await self._should_stop(should_stop):
-                                    await queue.put(
-                                        ("event", self._stopped_event(agent_run_id))
-                                    )
-                                    return
-                            final_message = self._final_output_message(
-                                output=node.data.output,
-                                tool_name=node.data.tool_name,
-                                tool_call_id=node.data.tool_call_id,
-                            )
-                            if final_message is not None:
-                                await queue.put(
-                                    (
-                                        "event",
-                                        AgentEvent(
-                                            type=AgentEventType.MESSAGE,
-                                            data=final_message,
-                                            agent_run_id=agent_run_id,
-                                        ),
-                                    )
-                                )
-                                if await self._should_stop(should_stop):
-                                    await queue.put(
-                                        ("event", self._stopped_event(agent_run_id))
-                                    )
-                                    return
-
-                        elif options.output_type is not None:
-                            final_message = self._final_output_message(
-                                output=node.data.output,
-                                tool_name=None,
-                                tool_call_id=None,
-                            )
-                            if final_message is not None:
-                                await queue.put(
-                                    (
-                                        "event",
-                                        AgentEvent(
-                                            type=AgentEventType.MESSAGE,
-                                            data=final_message,
-                                            agent_run_id=agent_run_id,
-                                        ),
-                                    )
-                                )
-                                if await self._should_stop(should_stop):
-                                    await queue.put(
-                                        ("event", self._stopped_event(agent_run_id))
-                                    )
-                                    return
-
-                # Tokens burned by abandoned attempts are still billed by the
-                # provider, so they are carried forward rather than forgotten.
-                totals = _usage_totals(run.usage, carried_usage)
-                await queue.put(
-                    (
-                        "event",
-                        AgentEvent(
-                            type=AgentEventType.USAGE,
-                            data=AgentRunUsage(
-                                model_name=options.model_name,
-                                usage_kind="llm",
-                                input_tokens=totals["input_tokens"],
-                                output_tokens=totals["output_tokens"],
-                                request_count=totals["requests"],
-                                tool_call_count=totals["tool_calls"],
-                                metadata={
-                                    "cache_write_tokens": totals["cache_write_tokens"],
-                                    "cache_read_tokens": totals["cache_read_tokens"],
-                                    "input_audio_tokens": totals["input_audio_tokens"],
-                                    "output_audio_tokens": totals[
-                                        "output_audio_tokens"
-                                    ],
-                                },
-                            ),
-                            agent_run_id=agent_run_id,
-                        ),
-                    )
-                )
+        node_loop = NodeLoop(
+            run_context=_new_run_context,
+            queue=queue,
+            streamer=streamer,
+            options=options,
+            agent_run_id=agent_run_id,
+            conversation_id=ctx.conversation_id,
+            final_output_message=self._final_output_message,
+        )
 
         async def _drive() -> None:
             try:
                 await _drive_with_retry(
-                    _drive_once,
+                    node_loop.drive_once,
                     queue=queue,
                     max_attempts=max_stream_attempts,
                     stream_reset=lambda: self._stream_reset_event(agent_run_id),
-                    stopped=lambda: self._stopped_event(agent_run_id),
-                    should_stop=lambda: self._should_stop(should_stop),
+                    stopped=streamer.stopped_event,
+                    should_stop=streamer.stop_requested,
                 )
             except BaseException as exc:  # noqa: BLE001 — relayed to parent below
                 # Includes CancelledError from our own task.cancel(). pydantic's
@@ -751,399 +454,6 @@ class PydanticAIHarness:
         _reraise_driver_failure(
             pending_error,
             cancelled_by_us=cancelled_by_us,
-            agent_run_id=agent_run_id,
-        )
-
-    async def _stream_model_request(
-        self,
-        node,
-        run,
-        *,
-        agent_run_id: UUID,
-        malformed_tool_call_ids: set[str],
-        should_stop: StopChecker | None,
-    ) -> AsyncIterator[AgentEvent]:
-        parts = StreamingParts(
-            emit_tokens=self.emit_tokens,
-            malformed_tool_call_ids=malformed_tool_call_ids,
-        )
-
-        async with node.stream(run.ctx) as request_stream:
-            async for event in request_stream:
-                if isinstance(event, PartStartEvent):
-                    handler = self._stream_part_start(
-                        event,
-                        parts,
-                        agent_run_id=agent_run_id,
-                        should_stop=should_stop,
-                    )
-                elif isinstance(event, PartDeltaEvent):
-                    handler = self._stream_part_delta(
-                        event,
-                        parts,
-                        agent_run_id=agent_run_id,
-                        should_stop=should_stop,
-                    )
-                elif isinstance(event, PartEndEvent):
-                    handler = self._stream_part_end(
-                        event,
-                        parts,
-                        agent_run_id=agent_run_id,
-                        should_stop=should_stop,
-                    )
-                elif isinstance(event, FinalResultEvent):
-                    handler = self._stream_final_result(
-                        parts,
-                        agent_run_id=agent_run_id,
-                        should_stop=should_stop,
-                    )
-                else:
-                    continue
-                async for agent_event in handler:
-                    yield agent_event
-                    # Every early exit inside a handler is a stop check, and
-                    # it announces itself by emitting STOPPED. Reading that
-                    # here is what lets the handlers stay ordinary
-                    # generators instead of needing a second channel to end
-                    # the stream.
-                    if agent_event.type is AgentEventType.STOPPED:
-                        return
-
-        async for token_event in self._emit_token_chunks(
-            parts.drain_all_token_buffers(force=True),
-            agent_run_id=agent_run_id,
-            should_stop=should_stop,
-        ):
-            yield token_event
-
-        for part_index in sorted(parts.kinds):
-            part = parts.objects.get(part_index)
-            if part is None:
-                continue
-            message = parts.completed_part_message(
-                part=part,
-                part_kind=parts.kinds[part_index],
-                part_content=parts.contents.get(part_index),
-            )
-            if message is not None:
-                yield AgentEvent(
-                    type=AgentEventType.MESSAGE,
-                    data=message,
-                    agent_run_id=agent_run_id,
-                )
-                if await self._should_stop(should_stop):
-                    yield self._stopped_event(agent_run_id)
-                    return
-
-    async def _emit_token_chunks(
-        self,
-        chunks: list[dict[str, str]],
-        *,
-        agent_run_id: UUID,
-        should_stop: StopChecker | None,
-    ) -> AsyncIterator[AgentEvent]:
-        """Emit token events, giving a stop request a chance between each.
-
-        This shape appeared five times: a run cancelled mid-response should not
-        keep streaming to the end of a buffered chunk list, so every chunk is a
-        place the stream can end. Emitting STOPPED is how it ends -- the driver
-        in `_stream_model_request` returns on seeing one, which closes whichever
-        handler is suspended here.
-        """
-        for chunk in chunks:
-            yield AgentEvent(
-                type=AgentEventType.TOKEN,
-                data=chunk,
-                agent_run_id=agent_run_id,
-            )
-            if await self._should_stop(should_stop):
-                yield self._stopped_event(agent_run_id)
-                return
-
-    async def _stream_part_start(
-        self,
-        event,
-        parts: StreamingParts,
-        *,
-        agent_run_id: UUID,
-        should_stop: StopChecker | None,
-    ) -> AsyncIterator[AgentEvent]:
-        """A part has begun: record its kind and open its token stream.
-
-        Stops the stream by emitting STOPPED; the driver returns on seeing one.
-        """
-        parts.objects[event.index] = event.part
-        if isinstance(event.part, TextPart):
-            parts.kinds[event.index] = "text"
-            content = event.part.content or ""
-            parts.contents[event.index] = content
-            async for token_event in self._emit_token_chunks(
-                parts.append_token_text("text", content),
-                agent_run_id=agent_run_id,
-                should_stop=should_stop,
-            ):
-                yield token_event
-        elif isinstance(event.part, ThinkingPart):
-            parts.kinds[event.index] = "thinking"
-            content = event.part.content or ""
-            parts.contents[event.index] = content
-            async for token_event in self._emit_token_chunks(
-                parts.append_token_text("thinking", content),
-                agent_run_id=agent_run_id,
-                should_stop=should_stop,
-            ):
-                yield token_event
-        elif isinstance(event.part, ToolCallPart):
-            parts.kinds[event.index] = "tool_call"
-            parts.tool_names[event.index] = event.part.tool_name
-            for token_chunk in parts.start_tool_stream(
-                event.index,
-                event.part.tool_name,
-            ):
-                yield AgentEvent(
-                    type=AgentEventType.TOKEN,
-                    data=token_chunk,
-                    agent_run_id=agent_run_id,
-                )
-                if await self._should_stop(should_stop):
-                    yield self._stopped_event(agent_run_id)
-                    return
-            initial_args = _tool_call_args_text(event.part.args)
-            if initial_args:
-                parts.tool_stream_has_args.add(event.index)
-                async for token_event in self._emit_token_chunks(
-                    parts.append_token_text("tool", initial_args),
-                    agent_run_id=agent_run_id,
-                    should_stop=should_stop,
-                ):
-                    yield token_event
-
-    async def _stream_part_delta(
-        self,
-        event,
-        parts: StreamingParts,
-        *,
-        agent_run_id: UUID,
-        should_stop: StopChecker | None,
-    ) -> AsyncIterator[AgentEvent]:
-        """A part grew: append the delta to what it is accumulating.
-
-        Stops the stream by emitting STOPPED; the driver returns on seeing one.
-        """
-        if isinstance(event.delta, TextPartDelta):
-            parts.kinds[event.index] = "text"
-            content_delta = event.delta.content_delta or ""
-            parts.contents[event.index] = (
-                parts.contents.get(event.index, "") + content_delta
-            )
-            async for token_event in self._emit_token_chunks(
-                parts.append_token_text("text", content_delta),
-                agent_run_id=agent_run_id,
-                should_stop=should_stop,
-            ):
-                yield token_event
-        elif isinstance(event.delta, ThinkingPartDelta):
-            parts.kinds.setdefault(event.index, "thinking")
-            content_delta = getattr(event.delta, "content_delta", "") or ""
-            if content_delta:
-                parts.contents[event.index] = (
-                    parts.contents.get(event.index, "") + content_delta
-                )
-                for token_chunk in parts.append_token_text(
-                    "thinking",
-                    content_delta,
-                ):
-                    yield AgentEvent(
-                        type=AgentEventType.TOKEN,
-                        data=token_chunk,
-                        agent_run_id=agent_run_id,
-                    )
-                    if await self._should_stop(should_stop):
-                        yield self._stopped_event(agent_run_id)
-                        return
-        elif isinstance(event.delta, ToolCallPartDelta):
-            parts.kinds.setdefault(event.index, "tool_call")
-            if event.delta.tool_name_delta:
-                parts.tool_names[event.index] = (
-                    parts.tool_names.get(event.index, "") + event.delta.tool_name_delta
-                )
-            tool_delta = _tool_call_delta_text(event.delta)
-            if tool_delta:
-                for token_chunk in parts.start_tool_stream(
-                    event.index,
-                    parts.tool_names.get(event.index, ""),
-                ):
-                    yield AgentEvent(
-                        type=AgentEventType.TOKEN,
-                        data=token_chunk,
-                        agent_run_id=agent_run_id,
-                    )
-                    if await self._should_stop(should_stop):
-                        yield self._stopped_event(agent_run_id)
-                        return
-                parts.tool_stream_has_args.add(event.index)
-                async for token_event in self._emit_token_chunks(
-                    parts.append_token_text("tool", tool_delta),
-                    agent_run_id=agent_run_id,
-                    should_stop=should_stop,
-                ):
-                    yield token_event
-
-    async def _stream_part_end(
-        self,
-        event,
-        parts: StreamingParts,
-        *,
-        agent_run_id: UUID,
-        should_stop: StopChecker | None,
-    ) -> AsyncIterator[AgentEvent]:
-        """A part finished: drain its buffer and persist what it became.
-
-        Stops the stream by emitting STOPPED; the driver returns on seeing one.
-        """
-        part_kind = parts.kinds.pop(event.index, None)
-        part_content = parts.contents.pop(event.index, None)
-        parts.objects.pop(event.index, None)
-        if isinstance(event.part, ToolCallPart) or part_kind == "tool_call":
-            for token_chunk in parts.start_tool_stream(
-                event.index,
-                getattr(event.part, "tool_name", None)
-                or parts.tool_names.get(event.index, ""),
-            ):
-                yield AgentEvent(
-                    type=AgentEventType.TOKEN,
-                    data=token_chunk,
-                    agent_run_id=agent_run_id,
-                )
-                if await self._should_stop(should_stop):
-                    yield self._stopped_event(agent_run_id)
-                    return
-            if event.index not in parts.tool_stream_has_args:
-                final_args = _tool_call_args_text(getattr(event.part, "args", None))
-                for token_chunk in parts.append_token_text(
-                    "tool",
-                    final_args or "{}",
-                ):
-                    yield AgentEvent(
-                        type=AgentEventType.TOKEN,
-                        data=token_chunk,
-                        agent_run_id=agent_run_id,
-                    )
-                    if await self._should_stop(should_stop):
-                        yield self._stopped_event(agent_run_id)
-                        return
-            async for token_event in self._emit_token_chunks(
-                parts.append_token_text("tool", "}"),
-                agent_run_id=agent_run_id,
-                should_stop=should_stop,
-            ):
-                yield token_event
-            parts.forget_tool_part(event.index)
-        async for token_event in self._emit_token_chunks(
-            parts.drain_all_token_buffers(force=True),
-            agent_run_id=agent_run_id,
-            should_stop=should_stop,
-        ):
-            yield token_event
-        message = parts.completed_part_message(
-            part=event.part,
-            part_kind=part_kind,
-            part_content=part_content,
-        )
-        if message is not None:
-            yield AgentEvent(
-                type=AgentEventType.MESSAGE,
-                data=message,
-                agent_run_id=agent_run_id,
-            )
-            if await self._should_stop(should_stop):
-                yield self._stopped_event(agent_run_id)
-                return
-
-    async def _stream_final_result(
-        self,
-        parts: StreamingParts,
-        *,
-        agent_run_id: UUID,
-        should_stop: StopChecker | None,
-    ) -> AsyncIterator[AgentEvent]:
-        """The model reached its final result.
-
-        Stops the stream by emitting STOPPED; the driver returns on seeing one.
-        """
-        async for token_event in self._emit_token_chunks(
-            parts.drain_all_token_buffers(force=True),
-            agent_run_id=agent_run_id,
-            should_stop=should_stop,
-        ):
-            yield token_event
-
-    async def _stream_tool_calls(
-        self,
-        node,
-        run,
-        *,
-        conversation_id: UUID,
-        agent_run_id: UUID,
-        malformed_tool_call_ids: set[str],
-        emitted_tool_response_ids: set[str],
-        should_stop: StopChecker | None,
-    ) -> AsyncIterator[AgentEvent]:
-        async with node.stream(run.ctx) as handle_stream:
-            async for event in handle_stream:
-                if isinstance(event, FunctionToolCallEvent):
-                    continue
-                if isinstance(event, FunctionToolResultEvent):
-                    result_part = event.part
-                    if result_part.tool_call_id in malformed_tool_call_ids:
-                        logger.debug(
-                            "agent.pydantic_ai.skipping_tool_result_malformed_call.diagnostic",
-                            tool_call_id=result_part.tool_call_id,
-                        )
-                        continue
-                    tool_output = result_part.content
-                    if isinstance(tool_output, BinaryContent):
-                        tool_output = {
-                            "type": "binary_content",
-                            "media_type": tool_output.media_type,
-                            "size_bytes": len(tool_output.data)
-                            if tool_output.data
-                            else 0,
-                        }
-                    elif hasattr(tool_output, "model_dump"):
-                        tool_output = to_json_value(tool_output)
-                    else:
-                        tool_output = to_json_value(tool_output)
-
-                    emitted_tool_response_ids.add(result_part.tool_call_id)
-                    yield AgentEvent(
-                        type=AgentEventType.MESSAGE,
-                        data=MessageDraft.of_tool_return(
-                            tool_name=result_part.tool_name or "unknown_tool",
-                            tool_call_id=result_part.tool_call_id,
-                            tool_result=tool_output,
-                            metadata={
-                                "tool_name": result_part.tool_name or "unknown_tool"
-                            },
-                        ),
-                        agent_run_id=agent_run_id,
-                    )
-                    if await self._should_stop(should_stop):
-                        yield self._stopped_event(agent_run_id)
-                        return
-
-    async def _should_stop(self, should_stop: StopChecker | None) -> bool:
-        if should_stop is None:
-            return False
-        try:
-            return await should_stop()
-        except Exception:
-            return False
-
-    def _stopped_event(self, agent_run_id: UUID) -> AgentEvent:
-        return AgentEvent(
-            type=AgentEventType.STOPPED,
-            data={"reason": "stop_requested"},
             agent_run_id=agent_run_id,
         )
 
@@ -1228,66 +538,5 @@ def _runtime_profile_model(options: HarnessOptions):
     )
 
 
-def _tool_call_delta_text(delta: ToolCallPartDelta) -> str:
-    if not delta.args_delta:
-        return ""
-    return _tool_call_args_text(delta.args_delta)
-
-
-def _tool_call_args_text(args: object) -> str:
-    if args is None or args == "":
-        return ""
-    if isinstance(args, str):
-        return args
-    return json.dumps(to_json_value(args), default=str)
-
-
-def _usage_value(usage: object, field: str) -> int:
-    value = getattr(usage, field, 0)
-    if value is None:
-        return 0
-    try:
-        return int(value)
-    except TypeError, ValueError:
-        return 0
-
-
 # Every usage field the harness reports, in one place so a retry can carry all of
 # them forward without the accumulator and the emitter drifting apart.
-_USAGE_FIELDS = (
-    "input_tokens",
-    "output_tokens",
-    "requests",
-    "tool_calls",
-    "cache_write_tokens",
-    "cache_read_tokens",
-    "input_audio_tokens",
-    "output_audio_tokens",
-)
-
-
-def _accumulate_usage(carried: dict[str, int], usage: object) -> None:
-    """Fold an abandoned attempt's usage into the running total."""
-    if usage is None:
-        return
-    for field in _USAGE_FIELDS:
-        carried[field] = carried.get(field, 0) + _usage_value(usage, field)
-
-
-def _usage_totals(usage: object, carried: dict[str, int]) -> dict[str, int]:
-    """This attempt's usage plus everything earlier attempts already spent."""
-    return {
-        field: _usage_value(usage, field) + carried.get(field, 0)
-        for field in _USAGE_FIELDS
-    }
-
-
-def _retry_backoff(attempt: int) -> float:
-    """Backoff before re-entering the graph, with jitter.
-
-    Short by design: the user is watching a stalled response, and the failure we
-    retry is a dropped connection rather than a loaded server. A provider that
-    wants longer says so via ``Retry-After``, which takes precedence.
-    """
-    base = min(_RETRY_BACKOFF_CAP_SECONDS, 0.5 * (3 ** (attempt - 1)))
-    return base + random.uniform(0, base / 2)
