+++ lemma-backend/app/modules/agent/tests/unit/test_vision_modes.py
@@ -18,7 +18,8 @@ from pydantic_ai import ToolReturn
 
 from app.modules.agent.domain.vision import AgentVisionMode, resolve_vision_mode
 from app.modules.agent.tools import vision_delegation
-from app.modules.agent.tools.pod import pydantic_adapter as pod_adapter
+from app.modules.agent.tools.pod import pod_common
+from app.modules.agent.tools.pod import pod_file_tools as pod_files
 from app.modules.agent.tools.pod.models import ViewDocumentPagesRequest
 from app.modules.agent.tools.context import BaseAgentContext
 
@@ -87,12 +88,12 @@ def _pdf_services(monkeypatch):
         async def __aexit__(self, *exc):
             return False
 
-    monkeypatch.setattr(pod_adapter, "pod_services", lambda deps: _Ctx())
+    monkeypatch.setattr(pod_common, "pod_services", lambda deps: _Ctx())
 
     async def fake_url(storage, key, expires_seconds=None):
         return f"https://signed/{key}", None
 
-    monkeypatch.setattr(pod_adapter, "build_object_url", fake_url)
+    monkeypatch.setattr(pod_files, "build_object_url", fake_url)
 
 
 def _ctx(mode: AgentVisionMode) -> SimpleNamespace:
@@ -111,7 +112,7 @@ class TestPdfPagesRespectTheMode:
     async def test_direct_still_receives_the_page_images(self, monkeypatch) -> None:
         _pdf_services(monkeypatch)
 
-        result = await pod_adapter.pod_view_document_pages(
+        result = await pod_files.pod_view_document_pages(
             _ctx(AgentVisionMode.DIRECT),
             ViewDocumentPagesRequest(path="/pod/report.pdf", page_start=1, page_end=2),
         )
@@ -135,7 +136,7 @@ class TestPdfPagesRespectTheMode:
         # Delegation now lives in one shared module, so patch it there.
         monkeypatch.setattr(vision_delegation, "describe_images", fake_describe)
 
-        result = await pod_adapter.pod_view_document_pages(
+        result = await pod_files.pod_view_document_pages(
             _ctx(AgentVisionMode.DELEGATED),
             ViewDocumentPagesRequest(
                 path="/pod/report.pdf",
@@ -172,7 +173,7 @@ class TestPdfPagesRespectTheMode:
         monkeypatch.delenv("VISION_MODEL", raising=False)
         monkeypatch.setattr(agent_settings, "vision_model", None)
 
-        result = await pod_adapter.pod_view_document_pages(
+        result = await pod_files.pod_view_document_pages(
             _ctx(AgentVisionMode.UNAVAILABLE),
             ViewDocumentPagesRequest(path="/pod/report.pdf", page_start=1),
         )
